+++ src/test/org/banks/xml/test/XMLParserSAXTests.java
@@ -3,9 +3,9 @@ package org.banks.xml.test;
 import org.banks.xml.parser.exception.InvalidFileException;
 import org.banks.xml.parser.exception.InvalidXMLException;
 import org.banks.xml.parser.model.Bank;
-import org.banks.xml.parser.service.builder.XMLParserManager;
+import org.banks.xml.parser.service.manager.XMLParserManager;
 import org.banks.xml.parser.service.parser.impl.ParserType;
-import org.banks.xml.parser.service.builder.impl.XMLParserManagerImpl;
+import org.banks.xml.parser.service.manager.impl.XMLParserManagerImpl;
 import org.banks.xml.parser.service.parser.XMLParser;
 import org.banks.xml.parser.service.parser.impl.XMLParserSAX;
 import org.testng.Assert;
